모달 수정, 로고 수정
