+++ vite.config.js
@@ -26,6 +26,6 @@ export default defineConfig({
       // 'D:/KB_6th/final_project/Mozi/mozi-backend/src/main/webapp/resources',
       //'C:/KB-PJT/goal/mozi-backend/src/main/webapp/resources',
       //'C:/KB_fullstack/final_project/backend/0728/mozi-backend/src/main/webapp/resources',
-      'C:/KB_Fullstack/accountConnect/mozi_4_backend/src/main/webapp/resources',
+      'C:/KB_Fullstack/10_finalProject/mozi-backend/src/main/webapp/resources',
   },
 });
